requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
tqdm>=4.64.0
opencv-python>=4.8.0

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
scikit-learn>=1.3.0

# Metrics and evaluation
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
kaleido>=0.2.1  # For plotly static image export

# Explainability
pytorch-grad-cam>=1.4.8

# Backend API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Logging and monitoring
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Optional but recommended
mlflow>=2.5.0
jupyter>=1.0.0
ipykernel>=6.25.0
